Moved Context out of GetRateRequest into its own file and separate class called Sep38Context.
